+++ CHANGELOG.md
@@ -8,6 +8,11 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 ## Unreleased
 
 ### Added
+- Asset functionality for mutators and handlers.
+
+## [2.0.0-beta.6-2] - 2018-10-22
+
+### Added
 - Add windows/386 to binary gcs releases
 - TLS authentication and encryption for etcd client and peer communication.
 - Added a debug log message for interval timer initial offset.
@@ -21,6 +26,7 @@ sensu/slack-handler.
 - Removed the `Error` store and type.
 
 ### Changed
+- Changed sensu-agent's internal asset manager to use BoltDB.
 - Changed sensuctl title colour to use terminal's configured default for bold
 text.
 - The backend no longer forcibly binds to localhost.